+++ spec/rubocop/cop/workit/comittee_assert_schema_confirm_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
-  it "registers an offense when using `assert_schema_conform` " \
-     "with no argument" do
+  it "registers an offense and autocorrects when using `assert_schema_conform` " \
+     "with no argument and `have_http_status`" do
     expect_offense(<<~RUBY)
       it 'something' do
         subject
@@ -22,7 +22,21 @@ RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
     RUBY
   end
 
-  it "does not register an offense when using `assert_schema_conform` " \
+  it "registers an offense when using `assert_schema_conform` " \
+     "with no argument" do
+    expect_offense(<<~RUBY)
+      it 'something' do
+        subject
+        do_something
+        assert_schema_conform
+        ^^^^^^^^^^^^^^^^^^^^^ Pass expected response status code to check it against the corresponding schema explicitly.
+      end
+    RUBY
+
+    expect_no_corrections
+  end
+
+  it "does not register an offense and autocorrects when using `assert_schema_conform` " \
      "with http status code" do
     expect_no_offenses(<<~RUBY)
       it 'something' do
@@ -32,7 +46,7 @@ RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
     RUBY
   end
 
-  it "registers an offense when using `assert_response_schema_confirm` " \
+  it "registers an offense and autocorrects when using `assert_response_schema_confirm` " \
      "with no argument" do
     expect_offense(<<~RUBY)
       it 'something' do