+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- Fix a false negative for `Workit/ComitteeAssertSchemaConfirm`. ([@ydah])
+
 ## 0.4.1 - 2022-12-12
+
 - Add documentation. ([@ydah])
 
 ## 0.4.0 - 2022-12-12